feat(ability): adds serialization to ability score queries
