--- apps/api/src/app/ability-score/models/ability-score.graphql.ts
@@ -1,10 +1,0 @@
-import { AbilityScore } from '@tabletop-companion/api-interface';
-export class AbilityScoreDTO implements AbilityScore {
-  id!: string;
-
-  name!: string;
-
-  value!: number;
-
-  characterId!: string;
-}
