+++ apps/api/src/app/ability-score/models/index.ts
@@ -1,4 +1,4 @@
-export * from './ability-score.graphql';
+export * from './ability-score.dto';
 export * from './ability-score-id';
 export * from './ability-score-insert';
 export * from './ability-score-update';
